Misc type conversion warning trivial fixes #rb maxime.mercier #robomerge 5.0 #preflight 618aa174cd591b516590f10c
[CL 18107621 by Yoan StAmant in ue5-main branch]

+++ Engine/Plugins/Runtime/StateTree/Source/StateTreeModule/Private/StateTreeExecutionContext.cpp
@@ -625,7 +625,7 @@ FStateTreeTransitionResult FStateTreeExecutionContext::TriggerTransitions(FState
 	for (FStateTreeHandle Handle = CurrentStatus.State; Handle.IsValid(); Handle = StateTree->States[Handle.Index].Parent)
 	{
 		const FBakedStateTreeState& State = StateTree->States[Handle.Index];
-		for (uint32 i = 0; i < State.TransitionsNum; i++)
+		for (uint8 i = 0; i < State.TransitionsNum; i++)
 		{
 			// All transition conditions must pass
 			const int16 TransitionIndex = State.TransitionsBegin + i;
